+++ tests/expressions_window/tests.py
@@ -1157,28 +1157,3 @@ class WindowFunctionTests(TestCase):
                     )
                 )
             )
-
-    def test_invalid_type_end_row_range(self):
-        msg = "end argument must be a positive integer, zero, or None, but got 'a'."
-        with self.assertRaisesMessage(ValueError, msg):
-            list(
-                Employee.objects.annotate(
-                    test=Window(
-                        expression=Sum("salary"),
-                        frame=RowRange(end="a"),
-                    )
-                )
-            )
-
-    def test_invalid_type_start_row_range(self):
-        msg = "start argument must be a negative integer, zero, or None, but got 'a'."
-        with self.assertRaisesMessage(ValueError, msg):
-            list(
-                Employee.objects.annotate(
-                    test=Window(
-                        expression=Sum("salary"),
-                        order_by=F("hire_date").asc(),
-                        frame=RowRange(start="a"),
-                    )
-                )
-            )