+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### unreleased
 
-- fix: #99 fix update value containing "where" cause exception.
+- fix: #99 update value containing "where" cause exception.
+- fix: #97 JSONField error in ClickHouse 24.8.
 
 ### 1.3.0